Create a mechanism for mapping API calls for database tables
The mechanism requires the developer to specify the structure of the
table for a specific API map (path). Each map has a name and a
description as well as a set of fields. These fields can be of
different types. Currently, there are three types of fields: for
numbers, strings and date-time values. All classes are still loaded
manually, but through a separate file called BayouCore.php in sys.