Remake the map mechanism since only one map can be loaded at a time
The ApiMaps class is no longer needed, because there is no need for
registering all maps during runtime. Instead, only the map that is
needed (requested in the API call) will be loaded into the $map object
to be delegated further request handling.

Each map file contains only the map of one database table. Therefore,
the original map was split into two separate files that correspond with
the apropriate tables in the project's demonstration database.