API calls must be made by sending a JSON structured request via POST
Check the request method type. If not POST, return the map description.
If the method is POST, read the raw POST data (sent JSON structure) and
convert it from JSON into a PHP object. If the POST data did not
contain a valid JSON or the JSON structure was not an object, it cannot
be considered a valid API request structure. If it is, send it to the
map for further parsing and handling. The map handler will, for now,
just echo back the contents of the JSON request structure while the
mechanism and the parser are being developed and tested.