Create a mechanism to get the number of affected rows after an execute
The execute method in the DataBase class will store the number of
affected rows to be available for retrieval once. If no execute was
ever performed, the method will return NULL and can return a proper
value 0 if no rows have been affected by the last execute method call.
Also, it will be set to NULL if the last execution ended with an error.

The BaseAction class will return 1 (always true in the WHERE clause) if
there are no action specification object parameters for it to parse.