Finish the add action handler
The add action handler requires the action specification object to
contain exactly two properties. One must be named find and the other
must be named values. The find parameter contains an action
specification object of its own which is parsed to give the WHERE
clause of the SQL query and the values parameter is an object with its
properties named after table fields of the requested map (table) and
their values are the new values to be set (modified) for all records
which are matched by the WHERE clause generated by parsing the action
specification object in the find parameter.

An example of the API request call for this action is:
{
  "edit": {
    "find" : {
      "_or" : [
        { "post_id" : { "_eq" : 3 } },
        { "user_id" : { "_ne" : 1 } }
      ]
    },
    "values" : {
      "user_id": 2,
      "title": "Edited via the API",
      "link": "edited-via-the-api",
      "content": "This post was added and then edited via the API"
    }
  }
}

This request will modify values of a record with the post_id 3 which is
not owned by the user with the user_id 1. The new values are given in
the properties of the object specified under the 'values' property.