Finish the remove action handler
The remove action handler's action specification object must have a
single property named find which is itself an action specification
object used for generating the WHERE clause of an SQL query when parsed
by the BaseAction parser mechanism.

Also, in this commit is an English syntax correction of a response
message in the edit action handler class.