The field type class must be able to reveal a database equivalent type
Created a method which each API map field type class must implement.
This method should return the name of the database type which is as
close as to or equivalent to the type this field is supposed to map.

This type name does not need to and most probably will not always
reflect the true implementation of the field it describes in the real
database table's field, but the type specified as the return value of
this method should be type which is possible for the implementation of
this field.