Change the return type of the description from string to an object
Including the name of the table, the map, a human readable description
and the API URL for the map. Also, lists all supported fields as well
as their name, API map key string, a human readable description of the
acceptable value (validation instruction) and the type as specified by
the database type equivalent method.