Move the API logic into its own class and remove all static resources
The LimitlessBayou class is a self contained application that handles
all logic previously ran from the project's index file.

All responses are created from within this new class instead of being
spawned everywhere around the project as ApiResponse instances.

The DataBase class was completely rewritten and is no longer a
wrapper with static methods for different tasks, but instead has to be
instanced to be used. All instancing and providing of this class object
is done by the main LimitlessBayou class and used by other classes.