Move DataBase access wrapper to a separate project
Created a composer package milantex/daw which now includes the DataBase
class previously written and used only in this project. This is done so
that the DataBase access wrapper class can be used for different
projects in the future. This project is the first one to use it.
Original DataBase.php file and the test file DataBase.test.php have
been removed from this project as they are not included in the vendor
directory through the composer package management mechanism.