Fix PHPUnit test cases for ApiResponse and ActionParameters classes
The removal of one Sys\ name space part was missed in the
ActionParameters test case class and the ApiResponse test case class is
fixed now that the API response only gets sent when the send method is
called. Before that, the output can be checked as it can be retrieved
from the ApiResponse clas object as a string at any time.